refactor: split invoker and router
